# docs/openapi.yaml
openapi: 3.0.3
info:
  title: sandboxd /v1 API
  version: "0.3.0"
  description: >
    The public, tenant-scoped API for sandboxd. This is the only surface a
    console or external integration should use — the internal /sandbox and
    /wake endpoints are not part of this contract. In the local Docker profile,
    requests may use a server session cookie, a bearer API token, or (when
    explicitly disabled) no credential. In the production AKS profile, requests
    use the server session created by OIDC authorization-code PKCE in the
    configured Entra tenant; local passwords and API keys are not production
    authentication mechanisms. Production ownership is keyed to the immutable
    Entra OID and roles are sandboxd.user or sandboxd.admin.
servers:
  - url: http://127.0.0.1:9090
    description: Default loopback bind (docker-compose maps SANDBOXD_API_BIND here)
security:
  - sessionCookieAuth: []
  - bearerAuth: []
  - {} # local auth-disabled mode only
...
paths:
...
  /v1/apps:
    get:
      operationId: listApps
      tags: [apps]
      summary: List apps (tenant-scoped)
      parameters:
        - in: query
          name: external_user_id
          schema: { type: string }
          description: Optional integration-tag filter.
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  apps:
                    type: array
                    items: { $ref: "#/components/schemas/App" }
    post:
      operationId: createApp
      tags: [apps]
      summary: Create an app
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/CreateAppRequest" }
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema: { $ref: "#/components/schemas/App" }
        "400": { $ref: "#/components/responses/Error" }
  /v1/apps/{id}:
    parameters:
      - $ref: "#/components/parameters/Id"
    get:
      operationId: getApp
      tags: [apps]
      summary: Get an app
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/App" }
        "404": { $ref: "#/components/responses/Error" }
    patch:
      operationId: updateApp
      tags: [apps]
      summary: Update an app (partial)
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/PatchAppRequest" }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/App" }
        "404": { $ref: "#/components/responses/Error" }
    delete:
      operationId: deleteApp
      tags: [apps]
      summary: Delete an app and EVERYTHING it owns — purges every sandbox (container, workspace image, snapshots) and all app rows (config, events, snapshots). Irreversible.
      responses:
        "204": { description: App and all its resources deleted }
        "404": { $ref: "#/components/responses/Error" }
  /v1/apps/{id}/sandbox:
    parameters:
      - $ref: "#/components/parameters/Id"
    post:
      operationId: createAppSandbox
      tags: [apps]
      summary: Create the app's current sandbox (one live sandbox per app)
      requestBody:
        required: false
        content:
          application/json:
            schema:
              type: object
              properties:
                template: { type: string, description: "Template name; omit for react-standard, 'blank' for empty." }
                runtime_preset: { type: string, description: "Runtime preset id (overrides the app default + template)." }
                ports: { type: array, items: { type: integer }, default: [3000] }
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Sandbox" }
        "404": { $ref: "#/components/responses/Error" }
        "409":
          description: The app already has a live sandbox.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
...
components:
  securitySchemes:
    sessionCookieAuth:
      type: apiKey
      in: cookie
      name: sbx_session
      description: >
        HttpOnly, host-only server session cookie. In production it is issued
        after the OIDC authorization-code PKCE flow; the cookie is not a bearer
        credential for preview hosts.
    bearerAuth:
      type: http
      scheme: bearer
  parameters:
    Id:
      in: path
      name: id
      required: true
      schema: { type: string }
  responses:
    Error:
      description: Error
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
  schemas:
...
    Error:
      type: object
      properties:
        error:
          type: object
          properties:
            code: { type: string }
            message: { type: string }
            retryable: { type: boolean }
    App:
      type: object
      properties:
        id: { type: string }
        name: { type: string }
        description: { type: string }
        tags: { type: array, items: { type: string } }
        external_user_id: { type: string }
        external_project_id: { type: string }
        runtime_preset: { type: string, description: "Default runtime preset id for new sandboxes." }
        current_sandbox_id: { type: string, description: "Empty when the app has no live sandbox." }
        created_at: { type: string, format: date-time }
        updated_at: { type: string, format: date-time }
...
    CreateAppRequest:
      type: object
      required: [name]
      properties:
        name: { type: string }
        description: { type: string }
        tags: { type: array, items: { type: string } }
        external_user_id: { type: string }
        external_project_id: { type: string }
        runtime_preset: { type: string, description: "Runtime preset id; see GET /v1/presets." }
        git:
          type: object
          description: >-
            Import from an HTTPS repo. PUBLIC repos (curated starters or any public URL)
            clone tokenless — omit credential_id. PRIVATE repos reference a token by
            credential_id (from /v1/git-credentials), never inline; it is decrypted
            control-plane-side at clone time and never enters the sandbox.
          required: [repo_url]
          properties:
            repo_url: { type: string, description: "HTTPS repo URL, tokenless (no credentials in the URL)." }
            branch: { type: string, description: "Branch to clone (default main)." }
            credential_id: { type: string, description: "Owner-scoped git_credential id. Omit for a public tokenless clone." }
    PatchAppRequest:
      type: object
      properties:
        name: { type: string }
        description: { type: string }
        tags: { type: array, items: { type: string } }
...
    Sandbox:
      type: object
      properties:
        id: { type: string }
        status: { type: string, enum: [creating, running, stopped, error] }
        template: { type: string }
        preview: { $ref: "#/components/schemas/Preview" }
        processes:
          type: array
          description: Supervised processes from the runtime manifest (web + workers).
          items: { $ref: "#/components/schemas/Process" }
        created_at: { type: string, format: date-time }
        updated_at: { type: string, format: date-time }
    Process:
      type: object
      properties:
        name: { type: string }
        kind: { type: string, enum: [web, worker] }
        running: { type: boolean }
        pid: { type: integer }
        restarts: { type: integer }
...
    Preview:
      type: object
      properties:
        url: { type: string, description: "Preview URL; the hostname port is the resolved web port (manifest/preset/3000)." }
        port: { type: integer, description: "Resolved web port the preview routes to (A1.5a)." }
        status: { type: string, enum: [starting, up, down, error] }
        last_checked_at: { type: string, format: date-time }
...